harness: extract native_tools dispatch/files/research/schedules/goals/sessions
- dispatch.rs: execute_tool* + global capability gating (~480 LoC)
- files.rs: file CRUD + helpers (~613 LoC)
- research.rs: fetch_url/web_search + SSRF/html (~308 LoC)
- schedules.rs: upsert_schedule + schedule_json (~156 LoC)
- goals.rs: goal CRUD + helpers (~264 LoC)
- sessions.rs: read_session_messages (~48 LoC)
- mod.rs shrinks 5978 -> 4249 LoC, avg ~690

All use super::* with pub(crate) re-exports, cargo check green.
