+++ src/harness/native_tools/mod.rs
@@ -27,12 +27,15 @@ use crate::harness::skills::registry::ToolRegistry;
 use crate::harness::skills::render::format_active_skill_context;
 
 mod common;
-mod sessions;
-mod goals;
-mod schedules;
-mod research;
-mod files;
-mod dispatch;
+pub mod files;
+pub mod research;
+pub mod schedules;
+pub mod goals;
+pub mod sessions;
+pub mod dispatch;
+pub use dispatch::{execute_tool, execute_tool_for_session, execute_tool_for_session_output};
+pub use files::{ensure_file_read_namespace, find_file_by_path, parse_file_uri};
+pub use goals::active_goals_context;
 mod registry;
 use common::{has_capability_action, normalize_logical_path, opt_str, opt_u64, opt_usize, req_str, require_capability, require_file_read, string_map, string_vec};
 pub use registry::{register_channel_tools, register_skill_tools, register_tools};
@@ -141,1154 +144,80 @@ pub(crate) fn artifact_uris_from_message_text(text: &str) -> Vec<String> {
 
 
 
-pub async fn execute_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    spec: &manifests::AgentSpec,
-    name: &str,
-    args: &Value,
-    config: &Config,
-) -> Result<Option<String>> {
-    execute_tool_for_session(
-        cp,
-        current_namespace,
-        current_agent,
-        "",
-        spec,
-        name,
-        args,
-        config,
-    )
-    .await
-}
-
-pub async fn execute_tool_for_session(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    spec: &manifests::AgentSpec,
-    name: &str,
-    args: &Value,
-    config: &Config,
-) -> Result<Option<String>> {
-    Ok(execute_tool_for_session_output(
-        cp,
-        current_namespace,
-        current_agent,
-        current_session,
-        spec,
-        name,
-        args,
-        config,
-    )
-    .await?
-    .map(|output| output.summary()))
-}
-
-pub async fn execute_tool_for_session_output(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    spec: &manifests::AgentSpec,
-    name: &str,
-    args: &Value,
-    config: &Config,
-) -> Result<Option<ToolOutput>> {
-    if let Some((capability, action)) = global_capability_for_tool(name) {
-        require_global_capability(config, capability, action)?;
-    }
-
-    if let Some(result) = artifact_tools::execute_output(
-        cp,
-        current_namespace,
-        current_agent,
-        current_session,
-        name,
-        args,
-    )
-    .await?
-    {
-        return Ok(Some(result));
-    }
-    if let Some(result) = a2a_tools::execute(
-        cp,
-        current_namespace,
-        current_agent,
-        current_session,
-        spec,
-        name,
-        args,
-    )
-    .await?
-    {
-        return Ok(Some(ToolOutput::text(result)));
-    }
-    if let Some(result) = task_tools::execute(
-        cp,
-        current_namespace,
-        current_agent,
-        current_session,
-        spec,
-        name,
-        args,
-    )
-    .await?
-    {
-        return Ok(Some(ToolOutput::text(result)));
-    }
-    // Keep the Monty execution future off the Tokio worker stack.  Its
-    // subprocess/lifecycle state is substantially larger than the generic
-    // native-tool dispatch future, and this branch is also evaluated for
-    // unrelated tools.
-    if let Some(result) = Box::pin(code_tools::execute(
-        config,
-        cp,
-        current_namespace,
-        current_agent,
-        current_session,
-        spec,
-        name,
-        args,
-    ))
-    .await?
-    {
-        return Ok(Some(result));
-    }
-
-    match name {
-        READ_SESSION_MESSAGES_TOOL => {
-            require_capability(spec, "sessions", "read:messages")?;
-            read_session_messages(cp, current_namespace, current_agent, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }        LIST_FILES_TOOL => {
-            require_file_read(spec)?;
-            list_files_tool(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        READ_FILE_TOOL => {
-            require_file_read(spec)?;
-            read_file_tool(cp, current_namespace, args).await.map(Some)
-        }
-        GET_FILE_METADATA_TOOL => {
-            require_file_read(spec)?;
-            get_file_metadata_tool(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        CREATE_FILE_TOOL => {
-            require_capability(spec, "files", "create")?;
-            create_file_tool(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        UPDATE_FILE_TOOL => {
-            require_capability(spec, "files", "update")?;
-            update_file_tool(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        DELETE_FILE_TOOL => {
-            require_capability(spec, "files", "delete")?;
-            delete_file_tool(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        FETCH_URL_TOOL => {
-            require_capability(spec, "research", "fetch_url")?;
-            fetch_url(args).await.map(ToolOutput::text).map(Some)
-        }
-        WEB_SEARCH_TOOL => {
-            require_capability(spec, "research", "web_search")?;
-            web_search(args).await.map(ToolOutput::text).map(Some)
-        }
-        ACTIVATE_SKILL_TOOL => {
-            let skill_name = req_str(args, "name")?;
-            let skills = namespace::load_available_skills(cp, current_namespace).await?;
-            let skill = namespace::find_effective_skill(&skills, skill_name)
-                .ok_or_else(|| anyhow!("skill '{}' is not available", skill_name))?;
-            let instructions = namespace::load_skill_instructions(cp, skill).await?;
-            if current_session.is_empty() {
-                return Ok(Some(ToolOutput::text(format_active_skill_context(&[(
-                    skill.clone(),
-                    instructions,
-                )]))));
-            }
-            crate::harness::sessions::activate_skill(
-                cp.kv.as_ref(),
-                current_namespace,
-                current_agent,
-                current_session,
-                skill_name,
-            )
-            .await?;
-            Ok(Some(ToolOutput::text(format!(
-                "Activated skill '{}'. Its workflow guidance is now active for this session.",
-                skill_name
-            ))))
-        }
-        DEACTIVATE_SKILL_TOOL => {
-            let skill_name = req_str(args, "name")?;
-            if current_session.is_empty() {
-                return Err(anyhow!("deactivate_skill requires a session"));
-            }
-            let removed = crate::harness::sessions::deactivate_skill(
-                cp.kv.as_ref(),
-                current_namespace,
-                current_agent,
-                current_session,
-                skill_name,
-            )
-            .await?;
-            Ok(Some(ToolOutput::text(if removed {
-                format!("Deactivated skill '{}'.", skill_name)
-            } else {
-                format!("Skill '{}' was not active.", skill_name)
-            })))
-        }
-        CHANNEL_PUBLISH_TOOL => {
-            let content = req_str(args, "content")?;
-            let message = crate::gateway::rpc::channels::publish_channel_message_from_session(
-                cp,
-                current_namespace,
-                current_agent,
-                current_session,
-                content,
-            )
-            .await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "published": true,
-                    "messageId": message.id,
-                    "channel": message.channel
-                }),
-            )?)))
-        }
-        CHANNEL_SKIP_REPLY_TOOL => {
-            let reason = opt_str(args, "reason").unwrap_or("");
-            crate::gateway::rpc::channels::skip_channel_reply_from_session(
-                cp,
-                current_namespace,
-                current_agent,
-                current_session,
-                reason,
-            )
-            .await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "published": false,
-                    "skipped": true
-                }),
-            )?)))
-        }
-        LIST_SCHEDULES_TOOL => {
-            require_capability(spec, "schedules", "inspect")?;
-            let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-            let agent = opt_str(args, "agent");
-            let enabled = args.get("enabled").and_then(Value::as_bool);
-            let limit = args.get("limit").and_then(Value::as_u64).unwrap_or(100) as usize;
-            let entries = cp
-                .kv
-                .list_entries(&keys::schedule_prefix(namespace), None)
-                .await?;
-            let mut schedules = Vec::new();
-            for (_key, value) in entries {
-                let schedule = resources_proto::Schedule::decode(value.as_slice())?;
-                let spec_model = schedule.spec.as_ref();
-                let matches_agent = agent
-                    .map(|target| {
-                        spec_model
-                            .and_then(|current| current.target.as_ref())
-                            .map(|target_model| target_model.agent == target)
-                            .unwrap_or(false)
-                    })
-                    .unwrap_or(true);
-                let matches_enabled = enabled
-                    .map(|value| {
-                        spec_model
-                            .map(|current| current.enabled == value)
-                            .unwrap_or(false)
-                    })
-                    .unwrap_or(true);
-                if matches_agent && matches_enabled {
-                    schedules.push(schedule_json(&schedule));
-                }
-                if schedules.len() >= limit {
-                    break;
-                }
-            }
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({ "schedules": schedules }),
-            )?)))
-        }
-        GET_SCHEDULE_TOOL => {
-            require_capability(spec, "schedules", "inspect")?;
-            let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-            let schedule_name = req_str(args, "name")?;
-            let schedule = cp
-                .kv
-                .get_msg::<resources_proto::Schedule>(&keys::schedule(namespace, schedule_name))
-                .await?
-                .ok_or_else(|| anyhow!("schedule '{}' not found", schedule_name))?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "schedule": schedule_json(&schedule)
-                }),
-            )?)))
-        }
-        CREATE_SCHEDULE_TOOL => {
-            require_capability(spec, "schedules", "create")?;
-            let schedule =
-                upsert_schedule(cp, current_namespace, current_agent, args, None).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "schedule": schedule_json(&schedule),
-                    "backendArmed": schedule.status.as_ref().map(|status| status.backend_armed).unwrap_or(false)
-                }),
-            )?)))
-        }
-        UPDATE_SCHEDULE_TOOL => {
-            require_capability(spec, "schedules", "update")?;
-            let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-            let schedule_name = req_str(args, "name")?;
-            let existing = cp
-                .kv
-                .get_msg::<resources_proto::Schedule>(&keys::schedule(namespace, schedule_name))
-                .await?
-                .ok_or_else(|| anyhow!("schedule '{}' not found", schedule_name))?;
-            let schedule =
-                upsert_schedule(cp, current_namespace, current_agent, args, Some(existing)).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "schedule": schedule_json(&schedule),
-                    "backendArmed": schedule.status.as_ref().map(|status| status.backend_armed).unwrap_or(false)
-                }),
-            )?)))
-        }
-        DELETE_SCHEDULE_TOOL => {
-            require_capability(spec, "schedules", "delete")?;
-            let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-            let schedule_name = req_str(args, "name")?;
-            let key = keys::schedule(namespace, schedule_name);
-            if let Some(schedule) = cp.kv.get_msg::<resources_proto::Schedule>(&key).await? {
-                if let Some(handle) = schedule.status.and_then(|status| status.backend_handle) {
-                    if let Err(error) = cp.scheduler.cancel(&handle).await {
-                        tracing::warn!(handle = %handle, error = %error, "failed to cancel schedule handle");
-                    }
-                }
-            }
-            cp.kv.delete(&key).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({ "success": true }),
-            )?)))
-        }
-        LIST_GOALS_TOOL => {
-            require_capability(spec, "goals", "inspect")?;
-            let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-            let agent = opt_str(args, "agent").unwrap_or(current_agent);
-            let session_id = opt_str(args, "session_id").unwrap_or(current_session);
-            let status_group = opt_str(args, "status_group");
-            let phase = opt_str(args, "phase");
-            let limit = args.get("limit").and_then(Value::as_u64).unwrap_or(100) as usize;
-            let goals = list_goals(cp, namespace, agent, session_id, status_group, phase, limit)
-                .await?
-                .into_iter()
-                .map(|goal| goal_json(&goal))
-                .collect::<Vec<_>>();
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goals": goals
-                }),
-            )?)))
-        }
-        GET_GOAL_TOOL => {
-            require_capability(spec, "goals", "inspect")?;
-            let goal =
-                get_goal_from_args(cp, current_namespace, current_agent, current_session, args)
-                    .await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goal": goal_json(&goal)
-                }),
-            )?)))
-        }
-        CREATE_GOAL_TOOL => {
-            require_capability(spec, "goals", "create")?;
-            let goal =
-                create_goal(cp, current_namespace, current_agent, current_session, args).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goal": goal_json(&goal)
-                }),
-            )?)))
-        }
-        UPDATE_GOAL_TOOL => {
-            require_capability(spec, "goals", "update")?;
-            let mut goal =
-                get_goal_from_args(cp, current_namespace, current_agent, current_session, args)
-                    .await?;
-            update_goal_from_args(&mut goal, args)?;
-            upsert_goal(cp, goal.clone()).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goal": goal_json(&goal)
-                }),
-            )?)))
-        }
-        COMPLETE_GOAL_TOOL => {
-            require_capability(spec, "goals", "update")?;
-            let mut goal =
-                get_goal_from_args(cp, current_namespace, current_agent, current_session, args)
-                    .await?;
-            let now = chrono::Utc::now().timestamp_micros();
-            goal.phase = crate::gateway::rpc::data_proto::GoalPhase::Succeeded as i32;
-            goal.updated_at = now;
-            goal.completed_at = now;
-            if let Some(summary) = opt_str(args, "progress_summary") {
-                goal.progress_summary = summary.to_string();
-            }
-            upsert_goal(cp, goal.clone()).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goal": goal_json(&goal)
-                }),
-            )?)))
-        }
-        BLOCK_GOAL_TOOL => {
-            require_capability(spec, "goals", "update")?;
-            let mut goal =
-                get_goal_from_args(cp, current_namespace, current_agent, current_session, args)
-                    .await?;
-            let now = chrono::Utc::now().timestamp_micros();
-            goal.phase = crate::gateway::rpc::data_proto::GoalPhase::Blocked as i32;
-            goal.updated_at = now;
-            goal.blocked_reason = req_str(args, "blocked_reason")?.to_string();
-            if let Some(summary) = opt_str(args, "progress_summary") {
-                goal.progress_summary = summary.to_string();
-            }
-            upsert_goal(cp, goal.clone()).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goal": goal_json(&goal)
-                }),
-            )?)))
-        }
-        _ => Ok(None),
-    }
-}
-
-fn require_global_capability(config: &Config, capability: &str, action: &str) -> Result<()> {
-    if global_capability_allowed(config, capability, action) {
-        Ok(())
-    } else {
-        Err(anyhow!(
-            "capability '{}:{}' is disabled by deployment configuration",
-            capability,
-            action
-        ))
-    }
-}
-
-fn global_capability_for_tool(name: &str) -> Option<(&'static str, &'static str)> {
-    match name {
-        READ_SESSION_MESSAGES_TOOL => Some(("sessions", "read:messages")),
-        LIST_FILES_TOOL | READ_FILE_TOOL | GET_FILE_METADATA_TOOL => Some(("files", "read")),
-        CREATE_FILE_TOOL => Some(("files", "create")),
-        UPDATE_FILE_TOOL => Some(("files", "update")),
-        DELETE_FILE_TOOL => Some(("files", "delete")),
-        FETCH_URL_TOOL => Some(("research", "fetch_url")),
-        WEB_SEARCH_TOOL => Some(("research", "web_search")),
-        LIST_SCHEDULES_TOOL | GET_SCHEDULE_TOOL => Some(("schedules", "inspect")),
-        CREATE_SCHEDULE_TOOL => Some(("schedules", "create")),
-        UPDATE_SCHEDULE_TOOL => Some(("schedules", "update")),
-        DELETE_SCHEDULE_TOOL => Some(("schedules", "delete")),
-        GET_GOAL_TOOL | LIST_GOALS_TOOL => Some(("goals", "inspect")),
-        CREATE_GOAL_TOOL => Some(("goals", "create")),
-        UPDATE_GOAL_TOOL | COMPLETE_GOAL_TOOL | BLOCK_GOAL_TOOL => Some(("goals", "update")),
-        _ => None,
-    }
-}
-
-async fn read_session_messages(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    args: &Value,
-) -> Result<String> {
-    let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-    let agent = opt_str(args, "agent").unwrap_or(current_agent);
-    let session_id = req_str(args, "session_id")?;
-    let limit = opt_usize(args, "limit").unwrap_or(20).clamp(1, 100);
-    let mut entries = cp
-        .kv
-        .list_entries(
-            &keys::session_message_prefix(namespace, agent, session_id),
-            Some(ListOptions::desc().limit(limit)),
-        )
-        .await?;
-    let mut messages = Vec::new();
-    entries.reverse();
-    for (_, bytes) in entries {
-        let message = data_proto::SessionMessage::decode(bytes.as_slice())?;
-        let role = data_proto::MessageRole::try_from(message.role)
-            .map(|role| format!("{role:?}"))
-            .unwrap_or_else(|_| message.role.to_string());
-        let text = message
-            .parts
-            .iter()
-            .filter(|part| part.part_type == data_proto::SessionMessagePartType::Text as i32)
-            .map(|part| part.content.as_str())
-            .collect::<String>();
-        messages.push(json!({
-            "id": message.id,
-            "role": role,
-            "text": text,
-            "createdAt": message.created_at,
-            "labels": message.labels,
-        }));
-    }
-    Ok(serde_json::to_string_pretty(&json!({
-        "namespace": namespace,
-        "agent": agent,
-        "sessionId": session_id,
-        "messages": messages,
-    }))?)
-}
+// dispatch moved to dispatch.rs
+// moved read_session_messages -> sessions.rs
 
 
 
 
 
-async fn list_files_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<String> {
-    let namespace = namespace_arg(current_namespace, args);
-    let prefix = opt_str(args, "prefix")
-        .map(normalize_logical_path)
-        .transpose()?;
-    let purpose = opt_str(args, "purpose")
-        .map(parse_file_purpose)
-        .transpose()?;
-    let index_policy = opt_str(args, "index_policy")
-        .map(parse_file_index_policy)
-        .transpose()?;
-    let limit = opt_usize(args, "limit").unwrap_or(50).clamp(1, 100);
-    let entries = list_files_by_filter(
-        cp,
-        &namespace,
-        prefix.as_deref().unwrap_or(""),
-        purpose,
-        index_policy,
-    )
-    .await?
-    .into_iter()
-    .take(limit)
-    .map(|file| file_json(&file, false))
-    .collect::<Vec<_>>();
-    Ok(serde_json::to_string_pretty(
-        &json!({ "entries": entries }),
-    )?)
-}
+// moved list_files_tool -> files.rs
 
-async fn read_file_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<ToolOutput> {
-    let (namespace, path) = file_location_from_args(current_namespace, args)?;
-    ensure_file_read_namespace(current_namespace, &namespace)?;
-    let file = find_file_by_path(cp, &namespace, &path)
-        .await?
-        .ok_or_else(|| anyhow!("File '{}' not found", path))?;
-    read_file_output(cp, &file).await
-}
+// moved read_file_tool -> files.rs
 
-async fn get_file_metadata_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<String> {
-    let (namespace, path) = file_location_from_args(current_namespace, args)?;
-    ensure_file_read_namespace(current_namespace, &namespace)?;
-    let file = find_file_by_path(cp, &namespace, &path)
-        .await?
-        .ok_or_else(|| anyhow!("File '{}' not found", path))?;
-    Ok(serde_json::to_string_pretty(&file_json(&file, false))?)
-}
+// moved get_file_metadata_tool -> files.rs
 
-async fn create_file_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<String> {
-    let (namespace, path) = file_location_from_args(current_namespace, args)?;
-    if find_file_by_path(cp, &namespace, &path).await?.is_some() {
-        return Err(anyhow!("File '{}' already exists", path));
-    }
-    let content = req_str(args, "content")?;
-    let media_type = opt_str(args, "media_type").unwrap_or("text/markdown");
-    let purpose = opt_str(args, "purpose")
-        .map(parse_file_purpose)
-        .transpose()?
-        .unwrap_or(resources_proto::FilePurpose::Artifact as i32);
-    let index_policy = opt_str(args, "index_policy")
-        .map(parse_file_index_policy)
-        .transpose()?
-        .unwrap_or(resources_proto::FileIndexPolicy::Search as i32);
-    let retention = opt_str(args, "retention")
-        .map(parse_file_retention)
-        .transpose()?
-        .unwrap_or(resources_proto::FileRetention::Retained as i32);
-    let file = upsert_file(
-        cp,
-        &namespace,
-        None,
-        &path,
-        media_type,
-        purpose,
-        index_policy,
-        retention,
-        content.as_bytes(),
-    )
-    .await?;
-    Ok(serde_json::to_string_pretty(&json!({
-        "file": file_json(&file, false),
-    }))?)
-}
+// moved create_file_tool -> files.rs
 
-async fn update_file_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<String> {
-    let (namespace, path) = file_location_from_args(current_namespace, args)?;
-    let existing = find_file_by_path(cp, &namespace, &path)
-        .await?
-        .ok_or_else(|| anyhow!("File '{}' not found", path))?;
-    let content = req_str(args, "content")?;
-    let existing_spec = existing.spec.as_ref();
-    let media_type = opt_str(args, "media_type")
-        .or_else(|| existing_spec.map(|spec| spec.media_type.as_str()))
-        .unwrap_or("text/markdown")
-        .to_string();
-    let purpose = opt_str(args, "purpose")
-        .map(parse_file_purpose)
-        .transpose()?
-        .or_else(|| existing_spec.map(|spec| spec.purpose))
-        .unwrap_or(resources_proto::FilePurpose::Artifact as i32);
-    let index_policy = opt_str(args, "index_policy")
-        .map(parse_file_index_policy)
-        .transpose()?
-        .or_else(|| existing_spec.map(|spec| spec.index_policy))
-        .unwrap_or(resources_proto::FileIndexPolicy::Search as i32);
-    let retention = opt_str(args, "retention")
-        .map(parse_file_retention)
-        .transpose()?
-        .or_else(|| existing_spec.map(|spec| spec.retention))
-        .unwrap_or(resources_proto::FileRetention::Retained as i32);
-    let file = upsert_file(
-        cp,
-        &namespace,
-        Some(existing),
-        &path,
-        &media_type,
-        purpose,
-        index_policy,
-        retention,
-        content.as_bytes(),
-    )
-    .await?;
-    Ok(serde_json::to_string_pretty(&json!({
-        "file": file_json(&file, false),
-    }))?)
-}
+// moved update_file_tool -> files.rs
 
-async fn delete_file_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<String> {
-    let (namespace, path) = file_location_from_args(current_namespace, args)?;
-    let file = find_file_by_path(cp, &namespace, &path)
-        .await?
-        .ok_or_else(|| anyhow!("File '{}' not found", path))?;
-    let name = file_name_from_file(&file);
-    let latest_key = crate::control::cas::latest_file_object_key(&namespace, &path);
-    let object_key = file
-        .status
-        .as_ref()
-        .and_then(|status| status.object_ref.as_ref())
-        .map(|object| object.key.clone());
-    let store = ResourceStore::new(cp.kv.clone(), cp.pubsub.clone());
-    let deleted = store.delete(&namespace, "File", &name).await?;
-    if deleted {
-        let cas = crate::control::cas::CasStore::new(cp.objects.clone());
-        if let Err(error) = cas.delete_object(&latest_key).await {
-            tracing::warn!(
-                error = %error,
-                object_key = %latest_key,
-                "failed to delete latest File object from native tool"
-            );
-        }
-        if let Some(object_key) = object_key {
-            if let Err(error) = cas.delete_object(&object_key).await {
-                tracing::warn!(
-                    error = %error,
-                    object_key = %object_key,
-                    "failed to delete File CAS object from native tool"
-                );
-            }
-        }
-    }
-    Ok(serde_json::to_string_pretty(&json!({
-        "deleted": deleted,
-        "namespace": namespace,
-        "path": path,
-    }))?)
-}
+// moved delete_file_tool -> files.rs
 
 
-async fn list_files_by_filter(
-    cp: &ControlPlane,
-    namespace: &str,
-    prefix: &str,
-    purpose: Option<i32>,
-    index_policy: Option<i32>,
-) -> Result<Vec<resources_proto::File>> {
-    let store = ResourceStore::new(cp.kv.clone(), cp.pubsub.clone());
-    let mut files = Vec::new();
-    for resource in store.list(namespace, Some("File")).await? {
-        let Some(file) = file_from_resource(resource) else {
-            continue;
-        };
-        let Some(spec) = file.spec.as_ref() else {
-            continue;
-        };
-        if purpose.is_some_and(|purpose| spec.purpose != purpose) {
-            continue;
-        }
-        if index_policy.is_some_and(|index_policy| spec.index_policy != index_policy) {
-            continue;
-        }
-        if !prefix.is_empty() && !spec.path.starts_with(prefix) {
-            continue;
-        }
-        files.push(file);
-    }
-    files.sort_by(|left, right| {
-        left.spec
-            .as_ref()
-            .map(|spec| spec.path.as_str())
-            .cmp(&right.spec.as_ref().map(|spec| spec.path.as_str()))
-    });
-    Ok(files)
-}
+// moved list_files_by_filter -> files.rs
 
 
-async fn find_file_by_path(
-    cp: &ControlPlane,
-    namespace: &str,
-    path: &str,
-) -> Result<Option<resources_proto::File>> {
-    Ok(list_files_by_filter(cp, namespace, path, None, None)
-        .await?
-        .into_iter()
-        .find(|file| file.spec.as_ref().map(|spec| spec.path.as_str()) == Some(path)))
-}
+// moved find_file_by_path -> files.rs
 
 struct ReadFileObject {
     object: crate::control::object_store::StoredObject,
     object_ref: data_proto::ObjectRef,
 }
 
-async fn read_file_content(cp: &ControlPlane, file: &resources_proto::File) -> Result<String> {
-    let read_object = read_file_object(cp, file).await?;
-    Ok(String::from_utf8_lossy(&read_object.object.bytes).to_string())
-}
+// moved read_file_content -> files.rs
 
-async fn read_file_output(cp: &ControlPlane, file: &resources_proto::File) -> Result<ToolOutput> {
-    let read_object = read_file_object(cp, file).await?;
-    let spec_media_type = file
-        .spec
-        .as_ref()
-        .map(|spec| spec.media_type.trim())
-        .unwrap_or_default();
-    let object_media_type = read_object.object.metadata.media_type.trim();
-    let media_type = if !spec_media_type.is_empty() {
-        spec_media_type.to_string()
-    } else if !object_media_type.is_empty() {
-        object_media_type.to_string()
-    } else {
-        file.spec
-            .as_ref()
-            .and_then(|spec| mime_guess::from_path(&spec.path).first_raw())
-            .unwrap_or("application/octet-stream")
-            .to_string()
-    };
-    let filename = {
-        let metadata_filename = read_object.object.metadata.filename.trim();
-        if !metadata_filename.is_empty() {
-            metadata_filename.to_string()
-        } else {
-            file.spec
-                .as_ref()
-                .map(|spec| {
-                    spec.path
-                        .trim_end_matches('/')
-                        .rsplit('/')
-                        .next()
-                        .unwrap_or_default()
-                        .to_string()
-                })
-                .unwrap_or_default()
-        }
-    };
-    let mut object_ref = read_object.object_ref;
-    object_ref.media_type = media_type.clone();
-    object_ref.filename = filename.clone();
-    Ok(ToolOutput::from_source_object(
-        read_object.object.bytes,
-        media_type,
-        filename,
-        object_ref,
-    ))
-}
+// moved read_file_output -> files.rs
 
-async fn read_file_object(
-    cp: &ControlPlane,
-    file: &resources_proto::File,
-) -> Result<ReadFileObject> {
-    let object_ref = file
-        .status
-        .as_ref()
-        .and_then(|status| status.object_ref.as_ref())
-        .ok_or_else(|| anyhow!("File has no objectRef"))?;
-    let object = crate::control::cas::CasStore::new(cp.objects.clone())
-        .get_object_decoded(&object_ref.key)
-        .await?
-        .ok_or_else(|| anyhow!("File object '{}' not found", object_ref.key))?;
-    let object_ref = crate::control::cas::object_ref_from_stored_object(&object_ref.key, &object);
-    Ok(ReadFileObject { object, object_ref })
-}
+// moved read_file_object -> files.rs
 
 
-async fn upsert_file(
-    cp: &ControlPlane,
-    namespace: &str,
-    existing: Option<resources_proto::File>,
-    path: &str,
-    media_type: &str,
-    purpose: i32,
-    index_policy: i32,
-    retention: i32,
-    content: &[u8],
-) -> Result<resources_proto::File> {
-    let store = ResourceStore::new(cp.kv.clone(), cp.pubsub.clone());
-    let name = existing
-        .as_ref()
-        .map(file_name_from_file)
-        .filter(|name| !name.is_empty())
-        .unwrap_or_else(|| keys::file_name_for_path(path));
-    let status = existing
-        .as_ref()
-        .and_then(|file| file.status.clone())
-        .unwrap_or_default();
-    let spec = resources_proto::FileSpec {
-        path: path.to_string(),
-        media_type: media_type.to_string(),
-        purpose,
-        index_policy,
-        retention,
-    };
-    let mut resource = store
-        .upsert(
-            namespace,
-            resource_model::file_resource(
-                namespace.to_string(),
-                name,
-                spec,
-                status,
-                file_resource_labels(purpose, index_policy, retention),
-            ),
-        )
-        .await?;
-    let uid = resource
-        .metadata
-        .as_ref()
-        .map(|metadata| metadata.uid.as_str())
-        .filter(|uid| !uid.is_empty())
-        .ok_or_else(|| anyhow!("File resource uid missing after upsert"))?;
-    let object_ref = write_file_objects(cp, namespace, uid, path, media_type, content).await?;
-    let status = resources_proto::FileStatus {
-        observed_generation: resource
-            .metadata
-            .as_ref()
-            .map(|metadata| metadata.generation)
-            .unwrap_or_default(),
-        phase: "Ready".to_string(),
-        conditions: Vec::new(),
-        object_ref: Some(object_ref),
-        updated_at: chrono::Utc::now().timestamp_micros(),
-        pending_upload: None,
-    };
-    resource.status = Some(resources_proto::ResourceStatus {
-        kind: Some(resources_proto::resource_status::Kind::File(status)),
-    });
-    let name = resource
-        .metadata
-        .as_ref()
-        .map(|metadata| metadata.name.clone())
-        .ok_or_else(|| anyhow!("File resource name missing"))?;
-    let resource = store
-        .patch_status(namespace, "File", &name, None, resource.status.unwrap())
-        .await?;
-    file_from_resource(resource).ok_or_else(|| anyhow!("invalid File resource"))
-}
+// moved upsert_file -> files.rs
 
-async fn write_file_objects(
-    cp: &ControlPlane,
-    namespace: &str,
-    file_uid: &str,
-    path: &str,
-    media_type: &str,
-    content: &[u8],
-) -> Result<resources_proto::FileObjectRef> {
-    let cas = crate::control::cas::CasStore::new(cp.objects.clone());
-    let object_ref = cas
-        .put_file(namespace, file_uid, path, content, media_type)
-        .await?;
-    cas.put_latest_file(namespace, path, content, media_type)
-        .await?;
-    Ok(resources_proto::FileObjectRef {
-        key: object_ref.key,
-        media_type: object_ref.media_type,
-        size_bytes: object_ref.size_bytes,
-        sha256: object_ref.sha256,
-        filename: object_ref.filename,
-        metadata: object_ref.metadata,
-    })
-}
+// moved write_file_objects -> files.rs
 
-fn file_from_resource(resource: resources_proto::Resource) -> Option<resources_proto::File> {
-    let spec = resource.spec.and_then(|spec| match spec.kind {
-        Some(resources_proto::resource_spec::Kind::File(spec)) => Some(spec),
-        _ => None,
-    })?;
-    let status = resource.status.and_then(|status| match status.kind {
-        Some(resources_proto::resource_status::Kind::File(status)) => Some(status),
-        _ => None,
-    });
-    Some(resources_proto::File {
-        metadata: resource.metadata,
-        spec: Some(spec),
-        status,
-    })
-}
+// moved file_from_resource -> files.rs
 
-fn file_name_from_file(file: &resources_proto::File) -> String {
-    file.metadata
-        .as_ref()
-        .map(|metadata| metadata.name.clone())
-        .unwrap_or_default()
-}
+// moved file_name_from_file -> files.rs
 
-fn file_json(file: &resources_proto::File, include_resource_name: bool) -> Value {
-    let namespace = file
-        .metadata
-        .as_ref()
-        .map(|metadata| metadata.namespace.as_str())
-        .unwrap_or_default();
-    let spec = file.spec.as_ref();
-    let path = spec.map(|spec| spec.path.as_str()).unwrap_or_default();
-    let object = file
-        .status
-        .as_ref()
-        .and_then(|status| status.object_ref.as_ref());
-    let mut value = json!({
-        "namespace": namespace,
-        "uri": file_uri(namespace, path),
-        "path": path,
-        "mediaType": spec.map(|spec| spec.media_type.as_str()).unwrap_or_default(),
-        "purpose": spec.map(|spec| file_purpose_label(spec.purpose)).unwrap_or("unspecified"),
-        "indexPolicy": spec.map(|spec| file_index_policy_label(spec.index_policy)).unwrap_or("unspecified"),
-        "retention": spec.map(|spec| file_retention_label(spec.retention)).unwrap_or("unspecified"),
-        "sizeBytes": object.map(|object| object.size_bytes).unwrap_or_default(),
-        "sha256": object.map(|object| object.sha256.as_str()).unwrap_or_default(),
-    });
-    if include_resource_name {
-        if let Some(map) = value.as_object_mut() {
-            map.insert("name".to_string(), json!(file_name_from_file(file)));
-        }
-    }
-    value
-}
+// moved file_json -> files.rs
 
 
-fn file_uri(namespace: &str, path: &str) -> String {
-    format!("file://{}{}", namespace, path)
-}
+// moved file_uri -> files.rs
 
-fn file_location_from_args(current_namespace: &str, args: &Value) -> Result<(String, String)> {
-    if let Some(uri) = opt_str(args, "uri") {
-        return parse_file_uri(uri);
-    }
-    let namespace = namespace_arg(current_namespace, args);
-    let path = normalize_logical_path(req_str(args, "path")?)?;
-    Ok((namespace, path))
-}
+// moved file_location_from_args -> files.rs
 
-fn namespace_arg(current_namespace: &str, args: &Value) -> String {
-    match opt_str(args, "namespace") {
-        Some("current" | "current_namespace" | "current namespace") | None => {
-            current_namespace.to_string()
-        }
-        Some(namespace) => namespace.to_string(),
-    }
-}
+// moved namespace_arg -> files.rs
 
-fn parse_file_uri(uri: &str) -> Result<(String, String)> {
-    let rest = uri
-        .trim()
-        .strip_prefix("file://")
-        .ok_or_else(|| anyhow!("file uri must start with 'file://'"))?;
-    let split = rest
-        .find('/')
-        .ok_or_else(|| anyhow!("file uri must include namespace and path"))?;
-    let namespace = rest[..split].trim();
-    if namespace.is_empty() {
-        return Err(anyhow!("file uri namespace is required"));
-    }
-    let path = normalize_logical_path(&rest[split..])?;
-    Ok((namespace.to_string(), path))
-}
+// moved parse_file_uri -> files.rs
 
-pub(super) fn ensure_file_read_namespace(
-    current_namespace: &str,
-    requested_namespace: &str,
-) -> Result<()> {
-    let requested_namespace = if requested_namespace == "current" {
-        current_namespace
-    } else {
-        requested_namespace
-    };
-    if crate::control::ns::ancestry(current_namespace)
-        .iter()
-        .any(|namespace| namespace == requested_namespace)
-    {
-        Ok(())
-    } else {
-        Err(anyhow!(
-            "file access to namespace '{}' is outside the current namespace ancestry",
-            requested_namespace
-        ))
-    }
-}
+// moved ensure_file_read_namespace -> files.rs
 
 
-fn file_resource_labels(
-    purpose: i32,
-    index_policy: i32,
-    retention: i32,
-) -> HashMap<String, String> {
-    HashMap::from([
-        (
-            "talon.impalasys.com/purpose".to_string(),
-            file_purpose_label(purpose).to_string(),
-        ),
-        (
-            "talon.impalasys.com/index-policy".to_string(),
-            file_index_policy_label(index_policy).to_string(),
-        ),
-        (
-            "talon.impalasys.com/retention".to_string(),
-            file_retention_label(retention).to_string(),
-        ),
-    ])
-}
+// moved file_resource_labels -> files.rs
 
-fn file_purpose_label(value: i32) -> &'static str {
-    match resources_proto::FilePurpose::try_from(value).ok() {
-        Some(resources_proto::FilePurpose::Memory) => "memory",
-        Some(resources_proto::FilePurpose::Artifact) => "artifact",
-        Some(resources_proto::FilePurpose::Skill) => "skill",
-        _ => "unspecified",
-    }
-}
+// moved file_purpose_label -> files.rs
 
-fn file_index_policy_label(value: i32) -> &'static str {
-    match resources_proto::FileIndexPolicy::try_from(value).ok() {
-        Some(resources_proto::FileIndexPolicy::None) => "none",
-        Some(resources_proto::FileIndexPolicy::Search) => "search",
-        Some(resources_proto::FileIndexPolicy::Retrieval) => "retrieval",
-        _ => "unspecified",
-    }
-}
+// moved file_index_policy_label -> files.rs
 
-fn file_retention_label(value: i32) -> &'static str {
-    match resources_proto::FileRetention::try_from(value).ok() {
-        Some(resources_proto::FileRetention::Retained) => "retained",
-        _ => "unspecified",
-    }
-}
+// moved file_retention_label -> files.rs
 
-fn parse_file_purpose(value: &str) -> Result<i32> {
-    let normalized = normalize_enum_input(value);
-    match normalized.as_str() {
-        "artifact" | "file-purpose-artifact" => Ok(resources_proto::FilePurpose::Artifact as i32),
-        "memory" | "file-purpose-memory" => Ok(resources_proto::FilePurpose::Memory as i32),
-        "skill" | "file-purpose-skill" => Ok(resources_proto::FilePurpose::Skill as i32),
-        _ => Err(anyhow!(
-            "unsupported File purpose '{}'; expected ARTIFACT, MEMORY, or SKILL",
-            value
-        )),
-    }
-}
+// moved parse_file_purpose -> files.rs
 
-fn parse_file_index_policy(value: &str) -> Result<i32> {
-    let normalized = normalize_enum_input(value);
-    match normalized.as_str() {
-        "none" | "file-index-policy-none" => Ok(resources_proto::FileIndexPolicy::None as i32),
-        "search" | "file-index-policy-search" => {
-            Ok(resources_proto::FileIndexPolicy::Search as i32)
-        }
-        "retrieval" | "file-index-policy-retrieval" => {
-            Ok(resources_proto::FileIndexPolicy::Retrieval as i32)
-        }
-        _ => Err(anyhow!(
-            "unsupported File index_policy '{}'; expected NONE, SEARCH, or RETRIEVAL",
-            value
-        )),
-    }
-}
+// moved parse_file_index_policy -> files.rs
 
-fn parse_file_retention(value: &str) -> Result<i32> {
-    let normalized = normalize_enum_input(value);
-    match normalized.as_str() {
-        "retained" | "file-retention-retained" => {
-            Ok(resources_proto::FileRetention::Retained as i32)
-        }
-        _ => Err(anyhow!(
-            "unsupported File retention '{}'; expected RETAINED",
-            value
-        )),
-    }
-}
+// moved parse_file_retention -> files.rs
 
-fn normalize_enum_input(value: &str) -> String {
-    value.trim().to_ascii_lowercase().replace(['_', ' '], "-")
-}
+// moved normalize_enum_input -> files.rs
 
 
 async fn create_artifact(
@@ -1585,235 +514,13 @@ async fn grant_artifact(
     }))?)
 }
 
-async fn fetch_url(args: &Value) -> Result<String> {
-    let url = req_str(args, "url")?;
-    let mut current_url = validate_public_http_url(url).await?;
-    let max_chars = args
-        .get("max_chars")
-        .and_then(Value::as_u64)
-        .unwrap_or(12_000)
-        .clamp(1_000, 40_000) as usize;
-    let mut redirects = 0usize;
-    let response = loop {
-        let client = http_client(&current_url)?;
-        let response = client.get(current_url.url.clone()).send().await?;
-        if !response.status().is_redirection() {
-            break response;
-        }
-        if redirects >= 5 {
-            return Err(anyhow!("too many redirects while fetching URL"));
-        }
-        let Some(location) = response.headers().get(reqwest::header::LOCATION) else {
-            break response;
-        };
-        let location = location
-            .to_str()
-            .map_err(|err| anyhow!("redirect Location is not valid UTF-8: {err}"))?;
-        current_url = validate_public_http_url(current_url.url.join(location)?.as_str()).await?;
-        redirects += 1;
-    };
-    let status = response.status().as_u16();
-    let final_url = response.url().to_string();
-    let content_type = response
-        .headers()
-        .get(reqwest::header::CONTENT_TYPE)
-        .and_then(|value| value.to_str().ok())
-        .unwrap_or("")
-        .to_string();
-    let body = response.text().await?;
-    let title = extract_title(&body);
-    let visible_text = compact_visible_text(&body, max_chars);
+// moved fetch_url -> research.rs
 
-    Ok(serde_json::to_string_pretty(&json!({
-        "url": url,
-        "finalUrl": final_url,
-        "status": status,
-        "contentType": content_type,
-        "title": title,
-        "text": visible_text,
-        "truncated": body.len() > max_chars
-    }))?)
-}
+// moved web_search -> research.rs
 
-async fn web_search(args: &Value) -> Result<String> {
-    let query = req_str(args, "query")?;
-    let limit = args
-        .get("limit")
-        .and_then(Value::as_u64)
-        .unwrap_or(5)
-        .clamp(1, 10) as usize;
-    let url = format!(
-        "https://duckduckgo.com/html/?q={}",
-        urlencoding::encode(query)
-    );
-    let url = validate_public_http_url(&url).await?;
-    let response = http_client(&url)?.get(url.url.clone()).send().await?;
-    let status = response.status().as_u16();
-    let body = response.text().await?;
-    let results = extract_duckduckgo_results(&body, limit);
+// moved upsert_schedule -> schedules.rs
 
-    Ok(serde_json::to_string_pretty(&json!({
-        "query": query,
-        "provider": "duckduckgo-html",
-        "status": status,
-        "results": results,
-        "citationPolicy": "Fetch a result URL with fetch_url before using it as evidence."
-    }))?)
-}
-
-async fn upsert_schedule(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    args: &Value,
-    existing: Option<resources_proto::Schedule>,
-) -> Result<resources_proto::Schedule> {
-    let namespace = opt_str(args, "namespace")
-        .unwrap_or(current_namespace)
-        .to_string();
-    let name = req_str(args, "name")?.to_string();
-    let existing_spec = existing
-        .as_ref()
-        .and_then(|schedule| schedule.spec.as_ref());
-    let existing_target = existing_spec.and_then(|spec| spec.target.as_ref());
-    let kind = scheduling::normalize_schedule_kind(
-        opt_str(args, "kind")
-            .or_else(|| existing_spec.map(|spec| spec.kind.as_str()))
-            .unwrap_or(""),
-    );
-    let cron = opt_str(args, "cron")
-        .map(str::to_string)
-        .or_else(|| existing_spec.map(|spec| spec.cron.clone()))
-        .unwrap_or_default();
-    let interval_seconds = args
-        .get("interval_seconds")
-        .and_then(Value::as_u64)
-        .map(|value| value as u32)
-        .or_else(|| existing_spec.map(|spec| spec.interval_seconds))
-        .unwrap_or_default();
-    let run_at = opt_str(args, "run_at")
-        .map(str::to_string)
-        .or_else(|| existing_spec.map(|spec| spec.run_at.clone()))
-        .unwrap_or_default();
-    let timezone = opt_str(args, "timezone")
-        .map(str::to_string)
-        .or_else(|| existing_spec.map(|spec| spec.timezone.clone()))
-        .unwrap_or_default();
-    let agent = opt_str(args, "agent")
-        .map(str::to_string)
-        .or_else(|| existing_target.map(|target| target.agent.clone()))
-        .unwrap_or_else(|| current_agent.to_string());
-    let session_mode = opt_str(args, "session_mode")
-        .map(str::to_string)
-        .or_else(|| existing_target.map(|target| target.session_mode.clone()))
-        .unwrap_or_else(|| "new".to_string());
-    let session_mode = scheduling::normalize_session_mode(&session_mode)?;
-    let session_id = opt_str(args, "session_id")
-        .map(str::to_string)
-        .or_else(|| existing_target.map(|target| target.session_id.clone()))
-        .unwrap_or_default();
-    let input_message = opt_str(args, "input_message")
-        .map(str::to_string)
-        .or_else(|| existing_spec.map(|spec| spec.input_message.clone()))
-        .unwrap_or_default();
-    let enabled = args
-        .get("enabled")
-        .and_then(Value::as_bool)
-        .or_else(|| existing_spec.map(|spec| spec.enabled))
-        .unwrap_or(true);
-    let labels = args
-        .get("labels")
-        .and_then(Value::as_object)
-        .map(|map| {
-            map.iter()
-                .filter_map(|(key, value)| {
-                    value
-                        .as_str()
-                        .map(|current| (key.clone(), current.to_string()))
-                })
-                .collect::<HashMap<_, _>>()
-        })
-        .or_else(|| existing.as_ref().map(|schedule| schedule.labels().clone()))
-        .unwrap_or_default();
-
-    let mut schedule = resource_model::schedule(
-        namespace.clone(),
-        name.clone(),
-        resources_proto::ScheduleSpec {
-            kind,
-            cron,
-            interval_seconds,
-            run_at,
-            timezone,
-            target: Some(resources_proto::ScheduleTarget {
-                agent,
-                workflow: String::new(),
-                session_mode,
-                session_id,
-            }),
-            input_message,
-            input_json: String::new(),
-            enabled,
-        },
-        existing
-            .and_then(|schedule| schedule.status)
-            .unwrap_or_default(),
-        labels,
-    );
-
-    scheduling::initialize_schedule(&mut schedule, chrono::Utc::now())?;
-    let next_run = schedule
-        .status
-        .as_ref()
-        .and_then(|status| status.next_run_at)
-        .and_then(chrono::DateTime::from_timestamp_micros);
-    scheduling::persist_schedule(cp.kv.as_ref(), &schedule).await?;
-    scheduling::arm_schedule(cp.scheduler.as_ref(), &mut schedule, next_run).await?;
-    scheduling::persist_schedule(cp.kv.as_ref(), &schedule).await?;
-    Ok(schedule)
-}
-
-fn schedule_json(schedule: &resources_proto::Schedule) -> Value {
-    let spec = schedule.spec.as_ref();
-    let status = schedule.status.as_ref();
-    let target = spec.and_then(|spec| spec.target.as_ref());
-    json!({
-        "name": schedule.name(),
-        "ns": schedule.namespace(),
-        "spec": {
-            "kind": spec.map(|spec| spec.kind.clone()).unwrap_or_default(),
-            "cron": spec.map(|spec| spec.cron.clone()).unwrap_or_default(),
-            "intervalSeconds": spec.map(|spec| spec.interval_seconds).unwrap_or_default(),
-            "runAt": spec.map(|spec| spec.run_at.clone()).unwrap_or_default(),
-            "timezone": spec.map(|spec| spec.timezone.clone()).unwrap_or_default(),
-            "target": {
-                "agent": target.map(|target| target.agent.clone()).unwrap_or_default(),
-                "sessionMode": target.map(|target| target.session_mode.clone()).unwrap_or_default(),
-                "sessionId": target.map(|target| target.session_id.clone()).unwrap_or_default(),
-            },
-            "inputMessage": spec.map(|spec| spec.input_message.clone()).unwrap_or_default(),
-            "enabled": spec.map(|spec| spec.enabled).unwrap_or(false),
-        },
-        "status": status.map(|status| json!({
-            "revision": status.revision,
-            "backendArmed": status.backend_armed,
-            "backendHandle": status.backend_handle,
-            "nextRunAt": status.next_run_at,
-            "lastRunAt": status.last_run_at,
-            "lastSessionId": status.last_session_id,
-            "lastError": status.last_error,
-            "claimedRunAt": status.claimed_run_at,
-            "claimExpiresAt": status.claim_expires_at,
-            "recentEvents": status.recent_events.iter().map(|event| json!({
-                "timestamp": event.timestamp,
-                "phase": event.phase,
-                "outcome": event.outcome,
-                "detail": event.detail,
-            })).collect::<Vec<_>>()
-        })).unwrap_or_else(|| json!({})),
-        "labels": schedule.labels(),
-    })
-}
+// moved schedule_json -> schedules.rs
 
 fn create_task(
     current_namespace: &str,
@@ -2383,267 +1090,35 @@ fn is_terminal_phase(value: i32) -> bool {
     )
 }
 
-async fn create_goal(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    args: &Value,
-) -> Result<data_proto::Goal> {
-    let objective = req_str(args, "objective")?.to_string();
-    let now = chrono::Utc::now().timestamp_micros();
-    let goal = data_proto::Goal {
-        id: crate::control::uuid::unique_name("goal"),
-        namespace: current_namespace.to_string(),
-        agent: current_agent.to_string(),
-        session_id: current_session.to_string(),
-        objective,
-        success_criteria: string_vec(args.get("success_criteria")),
-        phase: data_proto::GoalPhase::Running as i32,
-        progress_summary: opt_str(args, "progress_summary")
-            .unwrap_or("Goal created.")
-            .to_string(),
-        iteration: 0,
-        max_iterations: args
-            .get("max_iterations")
-            .and_then(Value::as_i64)
-            .unwrap_or_default()
-            .try_into()
-            .unwrap_or_default(),
-        created_at: now,
-        updated_at: now,
-        completed_at: 0,
-        blocked_reason: String::new(),
-        labels: string_map(args.get("labels")),
-        metadata: string_map(args.get("metadata")),
-    };
-    upsert_goal(cp, goal.clone()).await?;
-    Ok(goal)
-}
+// moved create_goal -> goals.rs
 
-async fn get_goal_from_args(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    args: &Value,
-) -> Result<data_proto::Goal> {
-    let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-    let agent = opt_str(args, "agent").unwrap_or(current_agent);
-    let session_id = opt_str(args, "session_id").unwrap_or(current_session);
-    let goal_id = req_str(args, "goal_id")?;
-    load_goal(cp, namespace, agent, session_id, goal_id)
-        .await?
-        .ok_or_else(|| anyhow!("goal '{}' not found", goal_id))
-}
+// moved get_goal_from_args -> goals.rs
 
-async fn load_goal(
-    cp: &ControlPlane,
-    namespace: &str,
-    agent: &str,
-    session_id: &str,
-    goal_id: &str,
-) -> Result<Option<data_proto::Goal>> {
-    cp.kv
-        .get_msg::<data_proto::Goal>(&keys::goal(namespace, agent, session_id, goal_id))
-        .await
-}
+// moved load_goal -> goals.rs
 
-async fn upsert_goal(cp: &ControlPlane, goal: data_proto::Goal) -> Result<()> {
-    cp.kv
-        .set_msg(
-            &keys::goal(&goal.namespace, &goal.agent, &goal.session_id, &goal.id),
-            &goal,
-        )
-        .await
-}
+// moved upsert_goal -> goals.rs
 
-async fn list_goals(
-    cp: &ControlPlane,
-    namespace: &str,
-    agent: &str,
-    session_id: &str,
-    status_group: Option<&str>,
-    phase: Option<&str>,
-    limit: usize,
-) -> Result<Vec<data_proto::Goal>> {
-    list_session_goals(cp, namespace, agent, session_id, status_group, phase, limit).await
-}
+// moved list_goals -> goals.rs
 
-async fn list_session_goals(
-    cp: &ControlPlane,
-    namespace: &str,
-    agent: &str,
-    session_id: &str,
-    status_group: Option<&str>,
-    phase: Option<&str>,
-    limit: usize,
-) -> Result<Vec<data_proto::Goal>> {
-    let mut goals = cp
-        .kv
-        .list_entries(&keys::goal_prefix(namespace, agent, session_id), None)
-        .await?
-        .into_iter()
-        .filter_map(|(_, value)| data_proto::Goal::decode(value.as_slice()).ok())
-        .filter(|goal| goal_matches(goal, status_group, phase))
-        .collect::<Vec<_>>();
-    goals.sort_by(|left, right| right.updated_at.cmp(&left.updated_at));
-    goals.truncate(limit);
-    Ok(goals)
-}
+// moved list_session_goals -> goals.rs
 
-fn goal_matches(goal: &data_proto::Goal, status_group: Option<&str>, phase: Option<&str>) -> bool {
-    if status_group
-        .is_some_and(|current| !goal_status_group(goal.phase).eq_ignore_ascii_case(current))
-    {
-        return false;
-    }
-    if phase.is_some_and(|current| parse_goal_phase(current).ok() != Some(goal.phase)) {
-        return false;
-    }
-    true
-}
+// moved goal_matches -> goals.rs
 
-fn update_goal_from_args(goal: &mut data_proto::Goal, args: &Value) -> Result<()> {
-    let now = chrono::Utc::now().timestamp_micros();
-    if let Some(phase) = opt_str(args, "phase") {
-        goal.phase = parse_goal_phase(phase)?;
-    }
-    if let Some(summary) = opt_str(args, "progress_summary") {
-        goal.progress_summary = summary.to_string();
-    }
-    if let Some(iteration) = args.get("iteration").and_then(Value::as_i64) {
-        goal.iteration = iteration.try_into().unwrap_or_default();
-    }
-    if let Some(blocked_reason) = opt_str(args, "blocked_reason") {
-        goal.blocked_reason = blocked_reason.to_string();
-    }
-    goal.updated_at = now;
-    if is_terminal_goal_phase(goal.phase) && goal.completed_at == 0 {
-        goal.completed_at = now;
-    }
-    Ok(())
-}
+// moved update_goal_from_args -> goals.rs
 
-fn goal_json(goal: &data_proto::Goal) -> Value {
-    json!({
-        "id": goal.id,
-        "namespace": goal.namespace,
-        "agent": goal.agent,
-        "sessionId": goal.session_id,
-        "objective": goal.objective,
-        "successCriteria": goal.success_criteria,
-        "phase": goal_phase_name(goal.phase),
-        "statusGroup": goal_status_group(goal.phase),
-        "progressSummary": goal.progress_summary,
-        "iteration": goal.iteration,
-        "maxIterations": goal.max_iterations,
-        "createdAt": goal.created_at,
-        "updatedAt": goal.updated_at,
-        "completedAt": goal.completed_at,
-        "blockedReason": goal.blocked_reason,
-        "labels": goal.labels,
-        "metadata": goal.metadata,
-    })
-}
+// moved goal_json -> goals.rs
 
-pub async fn active_goals_context(
-    cp: &ControlPlane,
-    namespace: &str,
-    agent: &str,
-    session_id: &str,
-) -> Result<Option<String>> {
-    let goals = list_goals(cp, namespace, agent, session_id, Some("active"), None, 20).await?;
-    if goals.is_empty() {
-        return Ok(None);
-    }
+// moved active_goals_context -> goals.rs
 
-    let mut lines = vec![
-        "# Active Talon Goals".to_string(),
-        "Keep these session-scoped objectives in view while deciding next steps.".to_string(),
-    ];
-    for goal in goals {
-        lines.push(format!(
-            "- {} [{}] {}",
-            goal.id,
-            goal_phase_name(goal.phase),
-            goal.objective
-        ));
-        if !goal.success_criteria.is_empty() {
-            lines.push(format!(
-                "  Success criteria: {}",
-                goal.success_criteria.join("; ")
-            ));
-        }
-        if !goal.progress_summary.is_empty() {
-            lines.push(format!("  Progress: {}", goal.progress_summary));
-        }
-        if !goal.blocked_reason.is_empty() {
-            lines.push(format!("  Blocked reason: {}", goal.blocked_reason));
-        }
-    }
-    Ok(Some(lines.join("\n")))
-}
+// moved parse_goal_phase -> goals.rs
 
-fn parse_goal_phase(value: &str) -> Result<i32> {
-    let phase = match value.trim().to_ascii_uppercase().as_str() {
-        "" | "UNSPECIFIED" => data_proto::GoalPhase::Unspecified,
-        "RUNNING" => data_proto::GoalPhase::Running,
-        "PAUSED" => data_proto::GoalPhase::Paused,
-        "NEEDS_REVIEW" | "NEEDS-REVIEW" => data_proto::GoalPhase::NeedsReview,
-        "SUCCEEDED" | "SUCCESS" | "COMPLETED" => data_proto::GoalPhase::Succeeded,
-        "FAILED" => data_proto::GoalPhase::Failed,
-        "BLOCKED" => data_proto::GoalPhase::Blocked,
-        "CANCELED" | "CANCELLED" => data_proto::GoalPhase::Canceled,
-        "EXPIRED" => data_proto::GoalPhase::Expired,
-        other => return Err(anyhow!("unsupported goal phase '{}'", other)),
-    };
-    Ok(phase as i32)
-}
+// moved goal_phase_name -> goals.rs
 
-fn goal_phase_name(value: i32) -> &'static str {
-    match data_proto::GoalPhase::try_from(value).ok() {
-        Some(data_proto::GoalPhase::Running) => "RUNNING",
-        Some(data_proto::GoalPhase::Paused) => "PAUSED",
-        Some(data_proto::GoalPhase::NeedsReview) => "NEEDS_REVIEW",
-        Some(data_proto::GoalPhase::Succeeded) => "SUCCEEDED",
-        Some(data_proto::GoalPhase::Failed) => "FAILED",
-        Some(data_proto::GoalPhase::Blocked) => "BLOCKED",
-        Some(data_proto::GoalPhase::Canceled) => "CANCELED",
-        Some(data_proto::GoalPhase::Expired) => "EXPIRED",
-        _ => "UNSPECIFIED",
-    }
-}
+// moved goal_status_group -> goals.rs
 
-fn goal_status_group(value: i32) -> &'static str {
-    if is_active_goal_phase(value) {
-        "ACTIVE"
-    } else if is_terminal_goal_phase(value) {
-        "TERMINAL"
-    } else {
-        "UNKNOWN"
-    }
-}
+// moved is_active_goal_phase -> goals.rs
 
-fn is_active_goal_phase(value: i32) -> bool {
-    matches!(
-        data_proto::GoalPhase::try_from(value).ok(),
-        Some(data_proto::GoalPhase::Running)
-            | Some(data_proto::GoalPhase::Paused)
-            | Some(data_proto::GoalPhase::NeedsReview)
-            | Some(data_proto::GoalPhase::Blocked)
-    )
-}
-
-fn is_terminal_goal_phase(value: i32) -> bool {
-    matches!(
-        data_proto::GoalPhase::try_from(value).ok(),
-        Some(data_proto::GoalPhase::Succeeded)
-            | Some(data_proto::GoalPhase::Failed)
-            | Some(data_proto::GoalPhase::Canceled)
-            | Some(data_proto::GoalPhase::Expired)
-    )
-}
+// moved is_terminal_goal_phase -> goals.rs
 
 
 
@@ -2741,235 +1216,31 @@ struct ValidatedHttpUrl {
     addrs: Vec<SocketAddr>,
 }
 
-fn http_client(target: &ValidatedHttpUrl) -> Result<reqwest::Client> {
-    let mut builder = reqwest::Client::builder()
-        .timeout(Duration::from_secs(20))
-        .user_agent("TalonResearchBot/0.1 (+https://impalasys.com)")
-        .redirect(reqwest::redirect::Policy::none());
-    if let Some(host) = target.host.as_deref() {
-        builder = builder.resolve_to_addrs(host, &target.addrs);
-    }
-    builder
-        .build()
-        .map_err(|err| anyhow!("research HTTP client failed to build: {err}"))
-}
+// moved http_client -> research.rs
 
-fn validate_http_url(value: &str) -> Result<()> {
-    let url = url::Url::parse(value).map_err(|err| anyhow!("invalid URL: {err}"))?;
-    match url.scheme() {
-        "http" | "https" => Ok(()),
-        scheme => Err(anyhow!("unsupported URL scheme '{}'", scheme)),
-    }
-}
+// moved validate_http_url -> research.rs
 
-async fn validate_public_http_url(value: &str) -> Result<ValidatedHttpUrl> {
-    validate_http_url(value)?;
-    let url = url::Url::parse(value).map_err(|err| anyhow!("invalid URL: {err}"))?;
-    let host = url
-        .host_str()
-        .ok_or_else(|| anyhow!("URL host is required"))?
-        .to_string();
-    let port = url
-        .port_or_known_default()
-        .ok_or_else(|| anyhow!("URL port is required"))?;
+// moved validate_public_http_url -> research.rs
 
-    if let Ok(ip) = host.parse::<IpAddr>() {
-        ensure_public_ip(ip)?;
-        return Ok(ValidatedHttpUrl {
-            url,
-            host: None,
-            addrs: vec![SocketAddr::new(ip, port)],
-        });
-    }
+// moved ensure_public_ip -> research.rs
 
-    let mut addrs = tokio::net::lookup_host((host.clone(), port))
-        .await
-        .map_err(|err| anyhow!("failed to resolve URL host '{host}': {err}"))?;
-    let mut public_addrs = Vec::new();
-    for addr in addrs.by_ref() {
-        ensure_public_ip(addr.ip())?;
-        public_addrs.push(addr);
-    }
-    if public_addrs.is_empty() {
-        return Err(anyhow!("URL host '{host}' resolved no addresses"));
-    }
-    Ok(ValidatedHttpUrl {
-        url,
-        host: Some(host),
-        addrs: public_addrs,
-    })
-}
+// moved is_blocked_ipv4 -> research.rs
 
-fn ensure_public_ip(ip: IpAddr) -> Result<()> {
-    let blocked = match ip {
-        IpAddr::V4(ip) => is_blocked_ipv4(ip),
-        IpAddr::V6(ip) => is_blocked_ipv6(ip),
-    };
-    if blocked {
-        Err(anyhow!("URL resolves to a non-public address"))
-    } else {
-        Ok(())
-    }
-}
+// moved is_blocked_ipv6 -> research.rs
 
-fn is_blocked_ipv4(ip: Ipv4Addr) -> bool {
-    let octets = ip.octets();
-    ip.is_private()
-        || ip.is_loopback()
-        || ip.is_link_local()
-        || ip.is_broadcast()
-        || ip.is_documentation()
-        || ip.is_unspecified()
-        || ip.is_multicast()
-        || octets[0] == 0
-        || octets[0] >= 224
-        || (octets[0] == 100 && (64..=127).contains(&octets[1]))
-        || (octets[0] == 169 && octets[1] == 254)
-}
+// moved extract_title -> research.rs
 
-fn is_blocked_ipv6(ip: Ipv6Addr) -> bool {
-    ip.is_loopback()
-        || ip.is_unspecified()
-        || ip.is_multicast()
-        || (ip.segments()[0] & 0xfe00) == 0xfc00
-        || (ip.segments()[0] & 0xffc0) == 0xfe80
-}
+// moved compact_visible_text -> research.rs
 
-fn extract_title(html: &str) -> String {
-    let lower = html.to_lowercase();
-    let Some(start) = lower.find("<title") else {
-        return String::new();
-    };
-    let Some(open_end) = lower[start..].find('>') else {
-        return String::new();
-    };
-    let content_start = start + open_end + 1;
-    let Some(close) = lower[content_start..].find("</title>") else {
-        return String::new();
-    };
-    decode_html_entities(&html[content_start..content_start + close])
-        .trim()
-        .to_string()
-}
+// moved remove_tag_blocks -> research.rs
 
-fn compact_visible_text(input: &str, max_chars: usize) -> String {
-    let without_scripts = remove_tag_blocks(input, "script");
-    let without_styles = remove_tag_blocks(&without_scripts, "style");
-    let mut text = String::with_capacity(without_styles.len().min(max_chars));
-    let mut in_tag = false;
-    let mut last_was_space = true;
-    for ch in without_styles.chars() {
-        match ch {
-            '<' => {
-                in_tag = true;
-                if !last_was_space {
-                    text.push(' ');
-                    last_was_space = true;
-                }
-            }
-            '>' => in_tag = false,
-            _ if in_tag => {}
-            _ if ch.is_whitespace() => {
-                if !last_was_space {
-                    text.push(' ');
-                    last_was_space = true;
-                }
-            }
-            _ => {
-                text.push(ch);
-                last_was_space = false;
-            }
-        }
-        if text.len() >= max_chars {
-            break;
-        }
-    }
-    decode_html_entities(text.trim())
-}
+// moved decode_html_entities -> research.rs
 
-fn remove_tag_blocks(input: &str, tag: &str) -> String {
-    let mut output = String::with_capacity(input.len());
-    let lower = input.to_lowercase();
-    let open_prefix = format!("<{}", tag);
-    let close = format!("</{}>", tag);
-    let mut pos = 0;
-    while let Some(start_rel) = lower[pos..].find(&open_prefix) {
-        let start = pos + start_rel;
-        output.push_str(&input[pos..start]);
-        if let Some(end_rel) = lower[start..].find(&close) {
-            pos = start + end_rel + close.len();
-        } else {
-            pos = input.len();
-            break;
-        }
-    }
-    output.push_str(&input[pos..]);
-    output
-}
+// moved extract_duckduckgo_results -> research.rs
 
-fn decode_html_entities(input: &str) -> String {
-    input
-        .replace("&amp;", "&")
-        .replace("&quot;", "\"")
-        .replace("&#39;", "'")
-        .replace("&apos;", "'")
-        .replace("&lt;", "<")
-        .replace("&gt;", ">")
-        .replace("&nbsp;", " ")
-}
+// moved extract_attr -> research.rs
 
-fn extract_duckduckgo_results(html: &str, limit: usize) -> Vec<Value> {
-    let mut results: Vec<Value> = Vec::new();
-    let mut pos = 0;
-    while results.len() < limit {
-        let Some(class_rel) = html[pos..].find("result__a") else {
-            break;
-        };
-        let class_pos = pos + class_rel;
-        let anchor_start = html[..class_pos].rfind("<a").unwrap_or(class_pos);
-        let Some(anchor_end_rel) = html[class_pos..].find("</a>") else {
-            break;
-        };
-        let anchor_end = class_pos + anchor_end_rel + "</a>".len();
-        let anchor = &html[anchor_start..anchor_end];
-        let Some(href) = extract_attr(anchor, "href") else {
-            pos = anchor_end;
-            continue;
-        };
-        let Some(url) = normalize_search_result_url(&href) else {
-            pos = anchor_end;
-            continue;
-        };
-        let title = compact_visible_text(anchor, 500);
-        if !title.is_empty() && !results.iter().any(|item| item["url"] == url) {
-            results.push(json!({
-                "title": title,
-                "url": url
-            }));
-        }
-        pos = anchor_end;
-    }
-    results
-}
-
-fn extract_attr(input: &str, attr: &str) -> Option<String> {
-    let needle = format!("{}=\"", attr);
-    let start = input.find(&needle)? + needle.len();
-    let end = input[start..].find('"')?;
-    Some(decode_html_entities(&input[start..start + end]))
-}
-
-fn normalize_search_result_url(href: &str) -> Option<String> {
-    if href.starts_with("http://") || href.starts_with("https://") {
-        return Some(href.to_string());
-    }
-    let query_start = href.find("uddg=")? + "uddg=".len();
-    let query = &href[query_start..];
-    let value = query.split('&').next().unwrap_or(query);
-    urlencoding::decode(value)
-        .ok()
-        .map(|value| value.into_owned())
-}
+// moved normalize_search_result_url -> research.rs
 
 fn default_access_expiry() -> i64 {
     access_expiry_from_ttl_seconds(24 * 60 * 60)
